+++ src/plaid/examples/dataset.py
@@ -9,7 +9,7 @@
 from plaid.containers.dataset import Dataset
 from datasets import load_dataset
 from plaid.bridges.huggingface_bridge import streamed_huggingface_dataset_to_plaid
-from plaid.examples import _HF_REPOS
+from plaid.examples.config import _HF_REPOS
 
 
 class _LazyDatasets: