+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- (plaid/examples) fix circular imports
+- (dataset.py) add optional `ids` argument to `from_list_of_samples`
+- (sample) add summarize and check_completeness functions
+- (dataset) add summarize_features and check_feature_completeness functions
 - (Hugging Face bridge) add datasetdict conversion, and simple function for plaid sample init from hf sample
 - (pipelines/plaid_blocks.py) add column transformer inverse_transform
 - Move to Muscat=2.5.0 (for tests and examples support)